move CurrentIndices to internal package
