+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/internal/CurrentIndicesFactory.kt
@@ -1,7 +1,5 @@
 package com.mapbox.navigation.base.internal
 
-import com.mapbox.navigation.base.CurrentIndices
-
 object CurrentIndicesFactory {
 
     fun createIndices(
