+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/internal/NavigationRouterV2.kt
@@ -1,7 +1,6 @@
 package com.mapbox.navigation.base.internal
 
 import com.mapbox.api.directions.v5.models.DirectionsRoute
-import com.mapbox.navigation.base.CurrentIndices
 import com.mapbox.navigation.base.route.NavigationRoute
 import com.mapbox.navigation.base.route.NavigationRouter
 import com.mapbox.navigation.base.route.NavigationRouterRefreshCallback
