+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/routerefresh/RefreshedRouteInfo.kt
@@ -1,6 +1,6 @@
 package com.mapbox.navigation.core.routerefresh
 
-import com.mapbox.navigation.base.CurrentIndices
+import com.mapbox.navigation.base.internal.CurrentIndices
 import com.mapbox.navigation.base.route.NavigationRoute
 
 internal data class RefreshedRouteInfo(
