+++ libnavigation-core/src/test/java/com/mapbox/navigation/core/CurrentIndicesProviderTest.kt
@@ -1,6 +1,6 @@
 package com.mapbox.navigation.core
 
-import com.mapbox.navigation.base.CurrentIndices
+import com.mapbox.navigation.base.internal.CurrentIndices
 import com.mapbox.navigation.base.internal.CurrentIndicesFactory
 import com.mapbox.navigation.base.trip.model.RouteProgress
 import com.mapbox.navigation.testing.MainCoroutineRule
